Fixed the active AI and width
It was actually too good. I unleashed a monster of epic proportions
that would in a small number of hands (between 300 and 400) could up to
quadruple its money. This mimics a human element by occasionally and
randomly betting twice. The width now auto adjusts on start-up to the
window width of the browser